+++ scripts/test_context_broker.sh
@@ -3,10 +3,7 @@
 echo "Obtaining token from Keycloak..."
 
 export token=`curl -d "client_id=client1&grant_type=password&username=admin&password=admin" -X POST --header "Host: keycloak:8080" 'http://localhost:8080/auth/realms/master/protocol/openid-connect/token' | \
-jq '.access_token'`
-
-export token="${token%\"}"
-export token="${token#\"}"
+jq -j '.access_token'`
 
 echo "Can I read all entities?"
 echo "==============================================================="